third_party/blink/web_tests/fast/events/key-events-in-editable-flexbox-expected.txt

PASS successfullyParsed is true

TEST COMPLETE

PASS targetDiv.innerText is "TEST"
PASS targetDiv.innerHTML is "<br>"
PASS targetDiv.innerText is "TEST"
TEST
